Move `aws_client_no_retry` fixture to the LocalStack testing fixtures
