+++ tests/conftest.py
@@ -1,6 +1,5 @@
 import os
 
-import botocore.config
 import pytest
 
 os.environ["LOCALSTACK_INTERNAL_TEST_RUN"] = "1"
@@ -104,19 +103,6 @@ def aws_client(aws_client_factory):
     from localstack.testing.aws.util import base_testing_aws_client
 
     return base_testing_aws_client(aws_client_factory)
-
-
-@pytest.fixture(scope="session")
-def aws_client_no_retry(aws_client_factory):
-    """
-    This fixture can be used to obtain Boto clients with disabled retries for testing.
-    botocore docs: https://boto3.amazonaws.com/v1/documentation/api/latest/guide/retries.html#configuring-a-retry-mode
-
-    Use this client when testing exceptions (i.e., with pytest.raises(...)) or expected errors (e.g., status code 500)
-    to avoid unnecessary retries and mitigate test flakiness if the tested error condition is time-bound.
-    """
-    no_retry_config = botocore.config.Config(retries={"max_attempts": 1})
-    return aws_client_factory(config=no_retry_config)
 
 
 @pytest.fixture(scope="session")
